+++ packages/node/src/Session.ts
@@ -40,6 +40,7 @@ import {
 } from "@inrupt/solid-client-authn-core";
 import { v4 } from "uuid";
 import EventEmitter from "events";
+import { exportJWK } from "jose";
 import type ClientAuthentication from "./ClientAuthentication";
 import { getClientAuthenticationWithDependencies } from "./dependencies";
 import IssuerConfigFetcher from "./login/oidc/IssuerConfigFetcher";
@@ -180,6 +181,14 @@ export class Session implements IHasSessionEventListener {
       refreshToken: sessionTokenSet.refreshToken,
       issuer: sessionTokenSet.issuer,
       tokenType: sessionTokenSet.dpopKey === undefined ? "Bearer" : "DPoP",
+      publicKey:
+        sessionTokenSet.dpopKey?.publicKey !== undefined
+          ? JSON.stringify(sessionTokenSet.dpopKey?.publicKey)
+          : undefined,
+      privateKey:
+        sessionTokenSet.dpopKey?.privateKey !== undefined
+          ? JSON.stringify(await exportJWK(sessionTokenSet.dpopKey?.privateKey))
+          : undefined,
     });
 
     const sessionInfo = await clientAuth.getSessionInfo(finalSessionId);